+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+
+* **New List Resource:** `mapbox_token` ([#99](https://github.com/DrFaust92/terraform-provider-mapbox/pull/99))
+
+ENHANCEMENTS:
+
+* resource/mapbox_token: Add resource identity (`token_id`, `username`) and support importing by identity ([#99](https://github.com/DrFaust92/terraform-provider-mapbox/pull/99))